# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SkillServe API",
        "description": "REST API for the SkillServe admin web system — a modular monolith (Laravel 13 + Sanctum). All responses use the standard envelope: { success, message, data, errors, meta }.",
        "contact": {
            "email": "group6@skillserve.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/administrators/{administrator}": {
            "get": {
                "tags": [
                    "Administrators"
                ],
                "summary": "Get an administrator",
                "description": "GET /api/administrators/{administrator} — a single administrator.",
                "operationId": "c688749a17a4a90be515af045438dd33",
                "parameters": [
                    {
                        "name": "administrator",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Administrator details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Request successful.",
                                    "data": {
                                        "id": 2,
                                        "first_name": "Jane",
                                        "last_name": "Doe",
                                        "name": "Jane Doe",
                                        "email": "jane.doe@skillserve.test",
                                        "status": "active",
                                        "roles": [
                                            "admin"
                                        ],
                                        "permissions": [
                                            "view reports"
                                        ],
                                        "last_login_at": "2026-08-07T09:30:00+00:00",
                                        "created_by": {
                                            "id": 1,
                                            "name": "System Administrator"
                                        },
                                        "created_at": "2026-08-07T08:00:00+00:00"
                                    },
                                    "errors": null,
                                    "meta": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated / expired token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Missing the manage administrators permission",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Administrator not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Administrators"
                ],
                "summary": "Update an administrator",
                "description": "PUT/PATCH /api/administrators/{administrator} — update name, email, role or status.",
                "operationId": "b0fa99b3e7752ac15f7b65b40c45bc3e",
                "parameters": [
                    {
                        "name": "administrator",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Jane"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane.doe@skillserve.test"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "admin"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "first_name": "Jane",
                                "last_name": "Doe",
                                "email": "jane.doe@skillserve.test",
                                "role": "admin",
                                "status": "active"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Administrator updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Administrator updated.",
                                    "data": {
                                        "id": 2,
                                        "first_name": "Jane",
                                        "last_name": "Doe",
                                        "name": "Jane Doe",
                                        "email": "jane.doe@skillserve.test",
                                        "status": "active",
                                        "roles": [
                                            "admin"
                                        ],
                                        "permissions": [
                                            "view reports"
                                        ],
                                        "last_login_at": "2026-08-07T09:30:00+00:00",
                                        "created_by": {
                                            "id": 1,
                                            "name": "System Administrator"
                                        },
                                        "created_at": "2026-08-07T08:00:00+00:00"
                                    },
                                    "errors": null,
                                    "meta": []
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated / expired token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Missing the manage administrators permission",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Administrator not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiEnvelope"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiEnvelope": {
                "required": [
                    "success",
                    "message"
                ],
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Request successful."
                    },
                    "data": {
                        "description": "Payload (object/array/null)",
                        "nullable": true
                    },
                    "errors": {
                        "description": "Field errors keyed by field",
                        "type": "object",
                        "nullable": true
                    },
                    "meta": {
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Paste the Sanctum token returned by POST /api/auth/login.",
                "bearerFormat": "sanctum-token",
                "scheme": "bearer"
            }
        }
    },
...
}